filter: add SetStringFromURL filter
This filter is needed for the FeedParser topology, but making it
more generic has no cost, so we end up with a filter that looks
for some user-defined strings in a record metadata URL key, such
as records read with the Files input. If one of the user-defined
strings is found in the URL, this new filter will sets an
user-defined field to that string. Records for which the URL
doesn't contain any of the strings are discarded.