Add MetricsClient.Close() method
Fast topology runs (< 1s) did not publish metrics since
baker.Main exited before they had the time.
This is also a problem in test.

Add a Close() method MetricsClient must implement, in which the
actual clients have the occasion to perform final flushes. We now
call this method in baker.Main().

Add a compile-time check on implementations of the MetricsClient
interface.