StatsDumper always dumps
Instead of early exiting if less than 1 second has elapsed since
the StatsDumper creation, we now always dump. A lot of useful
metrics are not time-related so it is too bad not dumping nothing
to avoid a division by zero when computing speed. Instead we
just print a 0 speed in that case, but dumps all other metrics.

Doing this also greatly simplifies testing StatsDumper output,
we don't need anymore to sleep for 1s in the tests, just to be
sure we'll get the StatsDumper output.