Add filter name metric tag (#200)
* Add and test makeUnivocal

* Store univocal filter names in topology

* Add univocal filter_name tag to filtered_lines metric

* Add MetricsClient.Close() method

Fast topology runs (< 1s) did not publish metrics since
baker.Main exited before they had the time.
This is also a problem in test.

Add a Close() method MetricsClient must implement, in which the
actual clients have the occasion to perform final flushes. We now
call this method in baker.Main().

Add a compile-time check on implementations of the MetricsClient
interface.

* go.{mod.sum}: bump datadog client from 3.2 to v5.1.0

* Comments

* Lint: unnecessary assigment

* StatsDumper always dumps

Instead of early exiting if less than 1 second has elapsed since
the StatsDumper creation, we now always dump. A lot of useful
metrics are not time-related so it is too bad not dumping nothing
to avoid a division by zero when computing speed. Instead we
just print a 0 speed in that case, but dumps all other metrics.

Doing this also greatly simplifies testing StatsDumper output,
we don't need anymore to sleep for 1s in the tests, just to be
sure we'll get the StatsDumper output.

* Rework/simplify StatsDumper tests

Also test the new feature: that filtered_lines metric is now
tagged by filter_name, and that filter names are univocal even
when the topology has multiple filters of the same type.

* Update CHANGELOG.md

* metrics/datadog: adapt tests to new datadog client version