Add and handle DropOnErrorDefault
FilterDesc.DropOnErrorDefault allows to set the default
DropOnError value for a filter. Dropping a record on error is
not the responsbility of the filter anymore since it can be
specified, on a per-filter basis, in the toml.

Allowing to set the default behaviour is useful for the many
filters which handled the dropping logic themselves. In this
way previously working topologies will continue to work without
any modification (i.e. without having to specific DropOnError=true).

So this will ease the upgrading process.