Topology keeps track of dropped records
Before, baker.Filter's were keeping track of dropped records and
publishing the number via Filter.Stats(), and StatsDumper was
doing the math.
Now that filters are not responsible for dropping records anymore
the topology keep track of per filter 'drop counter'.
StatsDumper keeps doing the math but it now uses topology as
source for dropped records in filters.